Move server implementation out of server program for testability
